talent endpoints

+++ apps/api/prisma/migrations/20231114005144_init/migration.sql
@@ -76,7 +76,7 @@ CREATE TABLE "TalentJobPreference" (
     "workMode" "WorkMode"[],
     "jobType" "JobType"[],
     "location" TEXT NOT NULL,
-    "expectedSalary" BIGINT NOT NULL,
+    "expectedSalary" INTEGER NOT NULL,
     "talentId" BIGINT NOT NULL,
     "createdAt" TIMESTAMP(3) NOT NULL DEFAULT CURRENT_TIMESTAMP,
     "updatedAt" TIMESTAMP(3) NOT NULL DEFAULT CURRENT_TIMESTAMP,
